+++ Source/Plugins/Plugin_VideoOGL/Src/TextureMngr.h
@@ -34,7 +34,7 @@ public:
         GLuint texture;
         u32 addr;
         u32 size_in_bytes;
-        u32 hash;
+        u64 hash;
         u32 paletteHash;
         u32 oldpixel; // used for simple cleanup
         TexMode0 mode; // current filter and clamp modes that texture is set to
